# [ROCm] Avoid using the default stream on ROCm as it is a performance killer