Fixing a ROCm performance regression added by #11744
Using the default 0 stream on ROCm hurts the performance. With this PR there is an uplift on tp=8:
LLama-3-70B
latency goes from 1.85s to 1.7s
LLama-2-70b bs=1 in=2048 out=128 scheduler-steps=10
latency goes from 1.8s to 1.58s